fix(templates): Set element in implicit bonds
Implicit bonds in residue templates did not account for element, so they
were missing in the bonded elements of the atom templates. Therefore,
the topological definition of the atom templates were incomplete, hence
template detection failed.

The fix required introducing a new notation for implicit bonds (not
backwards compatible). The elements of the implicit bonds are now
included in the residue template building.

BREAKING CHANGE: Change template notation for implicit bonds.