refactor!: Remove Topology class
The Topology class was meant to contain the functionality related to
topology, but it is essentially a Structure without coordinates, which
can be simulated by setting all coordinates to zero. Therefore, the code
was split into two namespaces without much improvement in usability or
code organization.

BREAKING CHANGE: Remove a type, remove instance variables, and change
method signatures.